cilium: convert fowarding_reason from int to uint8
Forwarding reason is a uint8 in trace and only holds a handful
of *CT definitions. Convert usage to uint8 to unify trace and
code.

After this its safe to & the encryption bit in.

+++ bpf/lib/conntrack.h
@@ -207,11 +207,11 @@ static inline bool __inline__ ct_entry_alive(const struct ct_entry *entry)
 	return !entry->rx_closing || !entry->tx_closing;
 }
 
-static inline int __inline__ __ct_lookup(void *map, struct __sk_buff *skb,
-					 void *tuple, int action, int dir,
-					 struct ct_state *ct_state,
-					 bool is_tcp, union tcp_flags seen_flags,
-					 __u32 *monitor)
+static inline __u8 __inline__ __ct_lookup(void *map, struct __sk_buff *skb,
+					  void *tuple, int action, int dir,
+					  struct ct_state *ct_state,
+					  bool is_tcp, union tcp_flags seen_flags,
+					  __u32 *monitor)
 {
 	struct ct_entry *entry;
 	int reopen;
